pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "suco-odoo-surface-mcp"
version = "0.1.0"
description = "User-equivalent Odoo access for AI agents via MCP"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["odoo", "mcp", "ai", "agent", "erp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "mcp>=1.27.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.urls]
Repository = "https://github.com/solutionsunity/odoo-surface-mcp"
Issues = "https://github.com/solutionsunity/odoo-surface-mcp/issues"

[project.scripts]
odoo-surface-mcp = "odoo_surface_mcp.main:main"

[tool.hatch.build.targets.wheel]
packages = ["odoo_surface_mcp"]
